Make cartouches work, very roughly
One bug these cartouches have for now is that they get stopped short when they encounter a diacritic. The right way to fix this would be to have the diacritics actually be treated as marks rather than zero-width characters, I think.

Also, they have hairline breaks which we may want to fix.